Final_WORKING_withoutTests

+++ YY6VGC_HSZF_2024251.Persistence.MsSql/RaceResultUpdate.cs
@@ -16,8 +16,10 @@ namespace YY6VGC_HSZF_2024251.Persistence.MsSql
         void AddNewDriver(Drivers driver);
         void DeleteDriver(int id);
         void UpdateCurrentDriverTeam(string driver, string newTeam);
-        void UpdateCurrentDriverPoints(string driver, int newPoint);
+        void UpdateCurrentDriverPoints(string driver);
         void AddDriverPoints(string[] namesToAddPoints);
+
+        void PointsAddDriver(string[] podiumDriverNameCollection);
     }
     public class RaceResultUpdate : IRaceResultUpdate
     {
@@ -30,6 +32,10 @@ namespace YY6VGC_HSZF_2024251.Persistence.MsSql
 
         public void AddNewDriver(Drivers driver)
         {
+            //var selfGrandPrixesId = 80;
+            driver.GrandPrixesId = 1;
+
+
             context.Drivers.Add(driver);
             context.SaveChanges();
         }
@@ -37,8 +43,16 @@ namespace YY6VGC_HSZF_2024251.Persistence.MsSql
         public void DeleteDriver(int id)
         {
             var driverToDelete = context.Drivers.FirstOrDefault(x => x.DriverId == id);
-            context.Drivers.Remove(driverToDelete);
-            context.SaveChanges();
+            if (driverToDelete == null)
+            {
+                Console.WriteLine("Nincs ilyen ID-val rendelkező versenyző az adatbázisban!");
+            }
+            else
+            {
+                context.Drivers.Remove(driverToDelete);
+                Console.WriteLine("Sikeres törlés!");
+            }
+            context.SaveChanges();           
 
         }
 
@@ -63,7 +77,15 @@ namespace YY6VGC_HSZF_2024251.Persistence.MsSql
         public void DeleteRace(int raceId)
         {
             var raceToDelete = context.GrandPrixes.FirstOrDefault(x => x.Id == raceId);
-            context.GrandPrixes.Remove(raceToDelete);
+            if (raceToDelete == null)
+            {
+                Console.WriteLine("Nem létezik ilyen ID-vel verseny az adatbázisban!");
+            }
+            else
+            {
+                context.GrandPrixes.Remove(raceToDelete);
+                Console.WriteLine("Sikeres törlés!");
+            }           
             context.SaveChanges();
         }
 
@@ -73,10 +95,9 @@ namespace YY6VGC_HSZF_2024251.Persistence.MsSql
             driverToUpdate.team = newTeam;
             context.SaveChanges();
         }
-        public void UpdateCurrentDriverPoints(string driver, int newPoint)
+        public void UpdateCurrentDriverPoints(string driver)
         {
-            var driverToUpdate = context.Drivers.FirstOrDefault(x => x.name == driver);
-            driverToUpdate.points += newPoint;
+            //driverToUpdate.points += newPoint;
             context.SaveChanges();
         }
         public void DeleteDriverPoints(string[] namesToDeletePoints)
@@ -142,7 +163,7 @@ namespace YY6VGC_HSZF_2024251.Persistence.MsSql
             findRaceId.Podium = newNames;
             context.SaveChanges();
             Console.WriteLine($"{findRaceId.Location} helyszínű verseny podiumja frissítve lett.");
-            
+
         }
 
         public void CheckOrCreatePilot(string pilotName, string[] teamname)
@@ -155,11 +176,30 @@ namespace YY6VGC_HSZF_2024251.Persistence.MsSql
             }
             else
             {
-                var newPilot = new Drivers { name = pilotName, team = "Unknown", points = 0};
+                var newPilot = new Drivers { name = pilotName, team = "Unknown", points = 0 };
                 context.Drivers.Add(newPilot);
                 context.SaveChanges();
                 Console.WriteLine($"A {newPilot} pilóta létrehozva az adatbázisban");
             }
+        }
+
+        public void PointsAddDriver(string[] podiumDriverNameCollection)
+        {
+            int[] points = new int[] { 25, 15, 8 }; // Pontszámok
+            for (int i = 0; i < podiumDriverNameCollection.Length; i++)
+            {
+                Drivers addNewDriver = new Drivers
+                {
+                    name = podiumDriverNameCollection[i],
+                    team = "Unknown",
+                    points = points[i], // A megfelelő pontot rendeljük hozzá
+                    nationality = "Unknown",
+                    GrandPrixesId = 1 // A versenyhez tartozó ID (használhatod az aktuális verseny ID-ját)
+                };
+                AddNewDriver(addNewDriver);
+            }
+
+
         }
     }
 }
